scheduler/requirements.txt
psycopg2==2.9.3
PyYAML==6.0.2
APScheduler==3.10.4
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.151.0
notion-client==2.2.1
